requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
